requirements.txt
autopep8==1.7.0
Babel==2.11.0
backcall==0.2.0
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
fastapi==0.85.1
jose==1.0.0
jsonschema==4.17.0
kafka-python==2.0.2
kiwisolver==1.4.4
lxml==4.9.1
mapbox-earcut==1.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
networkx==2.8.8
numpy==1.23.4
numpy-stl==2.17.1
opencv-python==4.6.0.66
pandas==1.5.1
passlib==1.7.4
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
python-utils==3.4.5
pytz==2022.2.1
requests==2.28.1
rsa==4.8
Rtree==1.0.1
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
Shapely==1.8.5.post1
sklearn==0.0.post1
sniffio==1.3.0
starlette==0.20.4
svg.path==6.2
svgwrite==1.4.3
trimesh==3.16.4
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
virtualenv==20.16.5
XlsxWriter==3.0.3
